Plotting WIP - Need to create separate Paper Figs folder
