+++ Demo_Lienard.py
@@ -3,10 +3,7 @@ import numpy as np
 
 from VISolver.Domains.Lienard import Lienard
 
-# from VISolver.Solvers.Euler_LEGS import Euler_LEGS
 from VISolver.Solvers.HeunEuler_LEGS import HeunEuler_LEGS
-# from VISolver.Solvers.AdamsBashforthEuler_LEGS import ABEuler_LEGS
-# from VISolver.Solvers.CashKarp_LEGS import CashKarp_LEGS
 
 from VISolver.Solver import Solve
 from VISolver.Options import (
@@ -20,22 +17,15 @@ from matplotlib import pyplot as plt
 
 def Demo():
 
-    #__CLOUD_SERVICES__##################################################
+    #__LIENARD_SYSTEM__##################################################
 
     # Define Network and Domain
     Domain = Lienard()
 
     # Set Method
-    # Method = Euler_LEGS(Domain=Domain)
     Method = HeunEuler_LEGS(Domain=Domain,Delta0=1e-5)
-    # Method = ABEuler_LEGS(Domain=Domain,Delta0=1e-4)
-    # Method = CashKarp_LEGS(Domain=Domain,Delta0=1e-6)
 
     # Initialize Starting Point
-    # Start = np.array([0,-1.0e-0])
-    # Start = np.array([0,-2.0e-0])
-    # Start = np.array([1.05151222,2.11233182])
-    # Start = np.array([0,0])
     Start = np.array([-.1,1.0])
 
     # Set Options
@@ -69,13 +59,6 @@ def Demo():
     ax.set_ylim([-2.5,2.5])
     ax.set_aspect('equal')
     plt.show()
-
-    # probs = [1/50] * 50
-    # np.random.multinomial(1, probs).argmax()
-    # np.random.choice(50, p=probs)
-    # grid = [np.array([.1,10.,6])]*5+[np.array([.1,1.,6])]*5
-    # grid = ListONP2NP(grid)
-    # Dinv = np.diag(grid[:,3])
 
 if __name__ == '__main__':
     Demo()
