+++ VISolver/Domains/CloudServices.py
@@ -1,14 +1,19 @@
 from __future__ import division
 import numpy as np
 from VISolver.Domain import Domain
+from numpy.polynomial.polynomial import polyval
+from scipy.optimize import minimize
+
+from matplotlib import pyplot as plt
 
 
 class CloudServices(Domain):
 
-    def __init__(self,Network,gap_alpha=2):
+    def __init__(self,Network,poly_splice=True,gap_alpha=2):
         self.UnpackNetwork(*Network)
         self.Network = (Network[0],Network[1])
         self.Dim = self.CalculateNetworkSize()
+        self.poly_splice = poly_splice
         self.gap_alpha = gap_alpha
 
     def F(self,Data):
@@ -33,107 +38,282 @@ class CloudServices(Domain):
         self.c_clouds = c_clouds
         self.H = H
         self.pref_bizes = pref_bizes
+        # self.coeff = np.array([3520,-4752,2564,-691,93,-5])
+        self.coeff = np.array([8,-20,26,-19,7,-1])
+        self.t0 = 1
+        self.tf = self.t0 + 1
 
     def CalculateNetworkSize(self):
         return 2*self.nClouds
+
+    def Nash(self,Data,rtol=1e-1,atol=1e-1):
+        isNash = np.zeros(self.nClouds,dtype=bool)
+        maxima = np.zeros((self.nClouds,2))
+        profits = np.zeros(self.nClouds)
+        success = np.zeros(self.nClouds,dtype=bool)
+        p = Data[:self.nClouds]
+        q = Data[self.nClouds:]
+        for i in xrange(self.nClouds):
+            pJ = np.sum(p)-p[i]
+            qJ = np.sum(q)-q[i]
+            x0 = np.array([p[i], q[i]])
+            Obj = lambda xi: -self.CloudProfit(i,xi,pJ,qJ)
+            Jac = lambda xi: -self.CloudJac(i,xi,pJ,qJ)
+            res = minimize(Obj, x0, method='L-BFGS-B', jac=Jac,
+                           bounds=[(0,None),(0,None)], tol=1e-8,
+                           options={'disp': False, 'maxiter': 1e5})
+            maxima[i,:] = res.x
+            profits[i] = -res.fun
+            success[i] = res.success
+            if np.allclose(res.x,x0,rtol,atol) and res.success:
+                isNash[i] = True
+        return isNash, maxima, profits, success
+
+    def Demand_iJ(self,i,pi,qi,pJ,qJ):
+        relprice = pi*self.nClouds/(pi+pJ)
+        relquali = qi*self.nClouds/(qi+qJ)
+
+        t = self.pref_bizes[:,i]*pi*qi*relprice*relquali
+
+        exp = self.H[:,i]*np.exp(-t**2)
+
+        if self.poly_splice:
+            poly = self.H[:,i]*np.exp(-self.t0**2)*polyval(t,self.coeff)
+
+            texp = (t <= self.t0)
+            tpoly = np.logical_and(t > self.t0,t < self.tf)
+
+            Qij = exp*texp + poly*tpoly
+        else:
+            Qij = exp
+
+        return Qij, t
+
+    def CloudProfit(self,i,xi,pJ,qJ):
+        pi, qi = xi
+        Q = self.Demand_iJ(i,pi,qi,pJ,qJ)[0]
+        Revenue = pi*Q
+        Cost = self.c_clouds[i]*Q*qi**(-2)
+        return np.sum(Revenue - Cost)
+
+    def CloudJac(self,i,xi,pJ,qJ):
+        pi, qi = xi
+        Qij, t = self.Demand_iJ(i,pi,qi,pJ,qJ)
+
+        c = self.c_clouds[i]
+        x = (pi-c/(qi**2))
+        a = 2*c/(qi**3)
+
+        dQ_dt_exp = -2*t*Qij
+
+        if self.poly_splice:
+            coeff_d1 = self.coeff[1:]*np.arange(1,len(self.coeff))
+            dQ_dt_poly = self.H[:,i]*np.exp(-self.t0**2)*polyval(t,coeff_d1)
+
+            texp = (t <= self.t0)
+            tpoly = np.logical_and(t > self.t0,t < self.tf)
+
+            dQ_dt = dQ_dt_exp*texp + dQ_dt_poly*tpoly
+        else:
+            dQ_dt = dQ_dt_exp
+
+        dt_dpj = t*(2/pi-1/(pi+pJ))
+        dt_dqj = t*(2/qi-1/(qi+qJ))
+
+        dpj = np.sum(Qij+x*(dQ_dt*dt_dpj))
+        dqj = np.sum(a*Qij+x*(dQ_dt*dt_dqj))
+
+        return np.array([dpj,dqj])
+
+    def CloudHes(self,i,xi,pJ,qJ):
+        pi, qi = xi
+        Qij, t = self.Demand_iJ(i,pi,qi,pJ,qJ)
+
+        c = self.c_clouds[i]
+        x = (pi-c/(qi**2))
+        a = 2*c/(qi**3)
+
+        dQ_dt_exp = -2*t*Qij
+        d2Q_dt2_exp = -2*(1-2*t**2)*Qij
+
+        if self.poly_splice:
+            coeff_d1 = self.coeff[1:]*np.arange(1,len(self.coeff))
+            dQ_dt_poly = self.H[:,i]*np.exp(-self.t0**2)*polyval(t,coeff_d1)
+            coeff_d2 = coeff_d1[1:]*np.arange(1,len(coeff_d1))
+            d2Q_dt2_poly = self.H[:,i]*np.exp(-self.t0**2)*polyval(t,coeff_d2)
+
+            texp = (t <= self.t0)
+            tpoly = np.logical_and(t > self.t0,t < self.tf)
+
+            dQ_dt = dQ_dt_exp*texp + dQ_dt_poly*tpoly
+            d2Q_dt2 = d2Q_dt2_exp*texp + d2Q_dt2_poly*tpoly
+        else:
+            dQ_dt = dQ_dt_exp
+            d2Q_dt2 = d2Q_dt2_exp
+
+        ps = pi+pJ
+        qs = qi+qJ
+
+        dt_dpj = t*(2/pi-1/ps)
+        dt_dqj = t*(2/qi-1/qs)
+
+        d2t_dpj2 = 2*t*(1/(pi**2)+1/(ps**2)-2/(pi*ps))
+        d2t_dqj2 = 2*t*(1/(qi**2)+1/(qs**2)-2/(qi*qs))
+
+        d2t_dpjdqj = t*(2/pi-1/ps)*(2/qi-1/qs)
+
+        dpj2 = 2*dQ_dt*dt_dpj+x*(dQ_dt*d2t_dpj2+d2Q_dt2*dt_dpj**2)
+        dqj2 = a*(2*dQ_dt*dt_dqj-3/qi*Qij)+x*(dQ_dt*d2t_dqj2+d2Q_dt2*dt_dqj**2)
+
+        dpjdqj = dQ_dt*(dt_dqj+a*dt_dpj) +\
+            x*(dQ_dt*d2t_dpjdqj+d2Q_dt2*dt_dpj*dt_dqj)
+
+        Jacobian = np.zeros((2,2))
+        Jacobian[0,0] = np.sum(dpj2)
+        Jacobian[0,1] = np.sum(dpjdqj)
+        Jacobian[1,0] = np.sum(dpjdqj)
+        Jacobian[1,1] = np.sum(dqj2)
+
+        return Jacobian
+
+    def Demand_ij(self,i,j,pi,qi,pJ,qJ):
+        relprice = pi*self.nClouds/(pi+pJ)
+        relquali = qi*self.nClouds/(qi+qJ)
+
+        t = self.pref_bizes[j,i]*pi*qi*relprice*relquali
+
+        exp = self.H[j,i]*np.exp(-t**2)
+
+        if self.poly_splice:
+            poly = self.H[j,i]*np.exp(-self.t0**2)*polyval(t,self.coeff)
+
+            texp = (t <= self.t0)
+            tpoly = np.logical_and(t > self.t0,t < self.tf)
+
+            Qij = exp*texp + poly*tpoly
+        else:
+            Qij = exp
+
+        return Qij, t
 
     def Demand_IJ(self,Data):
         p = Data[:self.nClouds]
         q = Data[self.nClouds:]
         relprice = p/np.mean(p)
         relquali = q/np.mean(q)
-        supply = p*q*relprice*relquali
-        market = self.pref_bizes*supply
-        # return self.H*np.exp(-market**2).clip(a_min=1e-6)
-        return self.H*np.exp(-market**2)
+
+        t = self.pref_bizes*p*q*relprice*relquali
+
+        exp = self.H*np.exp(-t**2)
+
+        if self.poly_splice:
+            poly = self.H*np.exp(-self.t0**2)*polyval(t,self.coeff)
+
+            texp = (t <= self.t0)
+            tpoly = np.logical_and(t > self.t0,t < self.tf)
+
+            Qij = exp*texp + poly*tpoly
+        else:
+            Qij = exp
+
+        return Qij, p, q, t
 
     def Demand(self,Data):
-        # why not just return np.sum(self.Demand_IJ(Data),axis=0)
-        p = Data[:self.nClouds]
-        q = Data[self.nClouds:]
-        relprice = p/np.mean(p)
-        relquali = q/np.mean(q)
-        supply = p*q*relprice*relquali
-        market = self.pref_bizes*supply
-        return np.sum(self.H*np.exp(-market**2),axis=0)
+        Qij, p, q = self.Demand_IJ(Data)[:3]
+        return np.sum(Qij,axis=0), p, q
 
     def CloudProfits(self,Data):
-        p = Data[:self.nClouds]
-        q = Data[self.nClouds:]
-        Q = self.Demand(Data)
+        Q, p, q = self.Demand(Data)[:3]
         Revenue = p*Q
         Cost = self.c_clouds*Q*q**(-2)
         return Revenue - Cost
 
     def dCloudProfits(self,Data):
-        p = Data[:self.nClouds]
-        q = Data[self.nClouds:]
-        relprice = p/np.mean(p)
-        relquali = q/np.mean(q)
-        supply = p*q*relprice*relquali
-        market = self.pref_bizes*supply
-        Qij = self.H*np.exp(-market**2)
-        # why not Qij = self.Demand_IJ(Data)
-        # because need p,q,and market (fij)
-
-        pfac = (2/p-1/np.sum(p))*2
-        qfac = (2/q-1/np.sum(q))*2
-        dfpj = -market**2*pfac
-        dfqj = -market**2*qfac
-
-        c = self.c_clouds
-
-        dpj = np.sum(Qij*(1+dfpj*(p-c*q**(-2))),axis=0)
-        dqj = np.sum(Qij*(2*c*q**(-3)+dfqj*(p-c*q**(-2))),axis=0)
-
-        return np.hstack([dpj,dqj])
-
-    def Jac(self,Data):
-
-        p = Data[:self.nClouds]
-        q = Data[self.nClouds:]
-
-        relprice = p/np.mean(p)
-        relquali = q/np.mean(q)
-        supply = p*q*relprice*relquali
-        market = self.pref_bizes*supply
-        fij = -market**2
-
-        Qij = self.H*np.exp(fij)
+        Qij, p, q, t = self.Demand_IJ(Data)
 
         ps = np.sum(p)
         qs = np.sum(q)
-
-        dfij_dpj = 2*fij*(2/p-1/ps)
-        dfij_dqj = 2*fij*(2/q-1/qs)
-        dfij_dpk = 2*fij*(-1/ps)  # same for every k
-        dfij_dqk = 2*fij*(-1/qs)  # same for every k
-
-        d2fij_dpj2 = 2*fij*(6/(p**2)-8/(p*ps)+3/(ps**2))
-        d2fij_dqj2 = 2*fij*(6/(q**2)-8/(q*qs)+3/(qs**2))
-
-        d2fij_dpjdqj = 2*fij*2*(2/q-1/qs)*(2/p-1/ps)
-
-        d2fij_dpjdpk = 2*fij*(-4/(p*ps)+3/(ps**2))
-        d2fij_dpjdqk = 2*fij*2*(-1/qs)*(2/p-1/ps)
-
-        d2fij_dqjdqk = 2*fij*(-4/(q*qs)+3/(qs**2))
-        d2fij_dqjdpk = 2*fij*2*(-1/ps)*(2/q-1/qs)
 
         c = self.c_clouds
         x = (p-c/(q**2))
         a = 2*c/(q**3)
 
-        dpjdpk = Qij*(dfij_dpk+(d2fij_dpjdpk+dfij_dpj*dfij_dpk)*x)
-        dpjdqk = Qij*(dfij_dqk+(d2fij_dpjdqk+dfij_dpj*dfij_dqk)*x)
+        dQ_dt_exp = -2*t*Qij
 
-        dqjdpk = Qij*(a*dfij_dpk+(d2fij_dqjdpk+dfij_dqj*dfij_dpk)*x)
-        dqjdqk = Qij*(a*dfij_dqk+(d2fij_dqjdqk+dfij_dqj*dfij_dqk)*x)
+        if self.poly_splice:
+            coeff_d1 = self.coeff[1:]*np.arange(1,len(self.coeff))
+            dQ_dt_poly = self.H*np.exp(-self.t0**2)*polyval(t,coeff_d1)
 
-        dpj2 = Qij*(2*dfij_dpj+(d2fij_dpj2+dfij_dpj**2)*x)
-        dqj2 = Qij*(2*a*dfij_dqj-3*a/q+(d2fij_dqj2+dfij_dqj**2)*x)
+            texp = (t <= self.t0)
+            tpoly = np.logical_and(t > self.t0,t < self.tf)
 
-        dpjdqj = Qij*(a*dfij_dpj+dfij_dqj+(d2fij_dpjdqj+dfij_dpj*dfij_dqj)*x)
+            dQ_dt = dQ_dt_exp*texp + dQ_dt_poly*tpoly
+        else:
+            dQ_dt = dQ_dt_exp
+
+        dt_dpj = t*(2/p-1/ps)
+        dt_dqj = t*(2/q-1/qs)
+
+        dpj = np.sum(Qij+x*(dQ_dt*dt_dpj),axis=0)
+        dqj = np.sum(a*Qij+x*(dQ_dt*dt_dqj),axis=0)
+
+        return np.hstack([dpj,dqj])
+
+    def Jac(self,Data):
+        Qij, p, q, t = self.Demand_IJ(Data)
+
+        ps = np.sum(p)
+        qs = np.sum(q)
+
+        c = self.c_clouds
+        x = (p-c/(q**2))
+        a = 2*c/(q**3)
+
+        dQ_dt_exp = -2*t*Qij
+        d2Q_dt2_exp = -2*(1-2*t**2)*Qij
+
+        if self.poly_splice:
+            coeff_d1 = self.coeff[1:]*np.arange(1,len(self.coeff))
+            dQ_dt_poly = self.H*np.exp(-self.t0**2)*polyval(t,coeff_d1)
+            coeff_d2 = coeff_d1[1:]*np.arange(1,len(coeff_d1))
+            d2Q_dt2_poly = self.H*np.exp(-self.t0**2)*polyval(t,coeff_d2)
+
+            texp = (t <= self.t0)
+            tpoly = np.logical_and(t > self.t0,t < self.tf)
+
+            dQ_dt = dQ_dt_exp*texp + dQ_dt_poly*tpoly
+            d2Q_dt2 = d2Q_dt2_exp*texp + d2Q_dt2_poly*tpoly
+        else:
+            dQ_dt = dQ_dt_exp
+            d2Q_dt2 = d2Q_dt2_exp
+
+        dt_dpj = t*(2/p-1/ps)
+        dt_dqj = t*(2/q-1/qs)
+        dt_dpk = t*(-1/ps)
+        dt_dqk = t*(-1/qs)
+
+        d2t_dpj2 = 2*t*(1/(p**2)+1/(ps**2)-2/(p*ps))
+        d2t_dqj2 = 2*t*(1/(q**2)+1/(qs**2)-2/(q*qs))
+
+        d2t_dpjdqj = t*(2/p-1/ps)*(2/q-1/qs)
+
+        d2t_dpjdpk = 2*t*(1/(ps**2)-1/(p*ps))
+        d2t_dpjdqk = t*(2/p-1/ps)*(-1/qs)
+
+        d2t_dqjdqk = 2*t*(1/(qs**2)-1/(q*qs))
+        d2t_dqjdpk = t*(2/q-1/qs)*(-1/ps)
+
+        dpj2 = 2*dQ_dt*dt_dpj+x*(dQ_dt*d2t_dpj2+d2Q_dt2*dt_dpj**2)
+        dqj2 = a*(2*dQ_dt*dt_dqj-3/q*Qij)+x*(dQ_dt*d2t_dqj2+d2Q_dt2*dt_dqj**2)
+
+        dpjdqj = dQ_dt*(dt_dqj+a*dt_dpj) +\
+            x*(dQ_dt*d2t_dpjdqj+d2Q_dt2*dt_dpj*dt_dqj)
+
+        dpjdpk = dQ_dt*dt_dpk+x*(dQ_dt*d2t_dpjdpk+d2Q_dt2*dt_dpj*dt_dpk)
+        dpjdqk = dQ_dt*dt_dqk+x*(dQ_dt*d2t_dpjdqk+d2Q_dt2*dt_dpj*dt_dqk)
+
+        dqjdqk = a*dQ_dt*dt_dqk+x*(dQ_dt*d2t_dqjdqk+d2Q_dt2*dt_dqj*dt_dqk)
+        dqjdpk = a*dQ_dt*dt_dpk+x*(dQ_dt*d2t_dqjdpk+d2Q_dt2*dt_dqj*dt_dpk)
 
         nc = self.nClouds
         Jacobian = np.zeros((2*nc,2*nc))
@@ -181,7 +361,6 @@ class CloudServices(Domain):
         return jac.transpose()
 
     def eig_stats(self,Data):
-
         jac = self.Jac(Data)
         eigs = np.real_if_close(np.linalg.eigvals(jac))
         eigs_r = np.real(eigs)
@@ -268,6 +447,40 @@ def CreateNetworkExample(ex=1):
                       [6,6,6,6,6],
                       [12,12,12,12,12]])
 
+    elif ex == 4:
+
+        # Cloud cost function coefficients
+        c_clouds = np.array([1.05,1.1,.95,1.15])
+
+        # Business preferences
+        pref_bizes = np.array([[.27,.27,.27,.38],
+                               [.34,.34,.34,.31],
+                               [.33,.33,.33,.26],
+                               [.25,.40,.40,.34]])
+
+        # Business scale factors
+        H = np.array([[11,11,11,11],
+                      [9,9,9,9],
+                      [6,6,6,6],
+                      [12,12,12,12]])
+
+    elif ex == 5:
+
+        # Cloud cost function coefficients
+        c_clouds = np.array([1.05,1.1,.95,1.15])
+
+        # Business preferences
+        pref_bizes = np.array([[.38,.38,.38,.27],
+                               [.34,.34,.34,.31],
+                               [.33,.33,.33,.26],
+                               [.25,.40,.40,.34]])
+
+        # Business scale factors
+        H = np.array([[11,11,11,11],
+                      [9,9,9,9],
+                      [6,6,6,6],
+                      [12,12,12,12]])
+
     else:
         raise NotImplementedError('There are only 3 predefined examples (1-3)')
 
@@ -278,7 +491,6 @@ def CreateNetworkExample(ex=1):
 
 
 def CreateRandomNetwork(nClouds=2,nBiz=2,seed=None):
-
     if seed is not None:
         np.random.seed(seed)
 
@@ -292,3 +504,102 @@ def CreateRandomNetwork(nClouds=2,nBiz=2,seed=None):
     H = np.tile(np.random.rand(nBiz)*10+5,(nClouds,1)).T
 
     return (nClouds,nBiz,c_clouds,H,pref_bizes)
+
+
+def plotDemandCurve(saveFig=False):
+    Network = CreateNetworkExample(ex=2)
+    Domain = CloudServices(Network=Network)
+    Domain2 = CloudServices(Network=Network,poly_splice=False)
+
+    i = 0
+    j = 3
+    pi = np.arange(0,.8,.001)
+    pJ = .01
+    qi = qJ = 1
+
+    Q = np.zeros_like(pi)
+    Q2 = Q.copy()
+    t = np.zeros_like(pi)
+    for p in xrange(len(pi)):
+        _Q, t = Domain.Demand_ij(i,j,pi[p],qi,pJ,qJ)
+        _Q2, _ = Domain2.Demand_ij(i,j,pi[p],qi,pJ,qJ)
+        Q[p] = _Q
+        Q2[p] = _Q2
+        t[p] = t
+
+    e = pi[np.argmax(t >= 1/np.sqrt(2))]
+    Qe = Q[np.argmax(t >= 1/np.sqrt(2))]/12.
+    p0 = pi[np.argmax(t >= Domain.t0)]
+    Qp0 = Q[np.argmax(t >= Domain.t0)]/12.
+    pf = pi[np.argmax(t >= Domain.tf)]
+    Qpf = Q[np.argmax(t >= Domain.tf)]/12.
+
+    fig = plt.figure()
+    ax = fig.add_subplot(111)
+
+    Qij, = ax.plot(pi,Q/12.,'k-', linewidth=5)
+    exp, = ax.plot(pi,Q2/12.,'k--', linewidth=5)
+    ax.plot(p0,Qp0,'ow',markersize=10)
+    ax.plot(pf,Qpf,'ow',markersize=10)
+    eps = .05
+    ax.plot([-eps,.8+eps],[0,0],'k-.')
+    ax.plot([0,0],[-eps,1+eps],'k-.')
+
+    ax.annotate('inelastic', xy=(.04, .97), xycoords='data',
+                xytext=(.1, .7), textcoords='data',
+                arrowprops=dict(arrowstyle='simple',facecolor='black'),
+                ha='center', va='center', size=18,
+                )
+    ax.annotate('elastic', xy=(e, Qe), xycoords='data',
+                xytext=(e+.2, Qe+.2), textcoords='data',
+                arrowprops=dict(frac=0.1,headwidth=10,width=4,
+                                facecolor='black',shrink=.1),
+                ha='center', va='center', size=18,
+                )
+    ax.annotate('splice', xy=(p0, Qp0), xycoords='data',
+                xytext=(p0+.2, Qp0+.2), textcoords='data',
+                arrowprops=dict(frac=0.1,headwidth=10,width=4,
+                                facecolor='black',shrink=.1),
+                ha='center', va='center', size=18,
+                )
+    ax.annotate('zero-cutoff', xy=(pf, Qpf+.02), xycoords='data',
+                xytext=(pf, Qpf+.2*np.sqrt(2)), textcoords='data',
+                arrowprops=dict(frac=0.1,headwidth=10,width=4,
+                                facecolor='black',shrink=.1),
+                ha='center', va='center', size=18,
+                )
+
+    ax.set_xlim(-eps,.8+eps)
+    ax.set_ylim(-eps,1+eps)
+
+    leg = plt.legend([Qij,exp], [r'$Q_{ij}^{}$', r'$H_{i}e^{-t_{ij}^2}$'],
+                     fontsize=20,fancybox=True)
+    plt.setp(leg.get_texts()[0], fontsize='20', va='center')
+    plt.setp(leg.get_texts()[1], fontsize='20', va='bottom')
+    plt.axis('off')
+
+    if saveFig:
+        plt.savefig("Demand.png",bbox_inches='tight')
+    return fig, ax
+
+
+def plotNonConcaveProfit(saveFig=False):
+    p = np.arange(0.01,20,.1)
+    Q11 = 10.*np.exp(-(p**2./(p+2.))**2.)
+    Q12 = np.exp(-(1./10.*p**2./(p+2.))**2.)
+    f = (Q11+Q12)*(p-1.)
+
+    fig = plt.figure()
+    fig.suptitle('Profit Non-Concave in Price',fontsize=18)
+    ax = fig.add_subplot(111)
+
+    ax.plot(p,f,'k',lw=2)
+
+    ax.set_xlim([0,20])
+    ax.set_ylim([-2,6])
+    plt.xlabel('$p_1$',fontsize=14)
+    plt.ylabel('$\pi_1$',fontsize=14)
+
+    if saveFig:
+        plt.savefig('NonConcave.png',bbox_inches='tight')
+    return fig, ax
